test: validate public action distribution
